1. First Input auto focus. 2. Separate file for tableStructure and selectOption 3. Style change to scss. 4. Some bug fixing.
