+++ src/_shared/selectOption.js
@@ -1,0 +1,9 @@
+export const selectOption = 
+[
+    { value: 'tape', displayValue: 'Tape' },
+    { value: 'cd', displayValue: 'CD' },
+    { value: 'good-day', displayValue: 'Good Day' },
+    { value: 'burbone', displayValue: 'Burbone' },
+    { value: 'hide&seek', displayValue: 'Hide&Seek' },
+    { value: 'parle', displayValue: 'Parle' },
+];
