+++ src/_shared/tableStructure.js
@@ -1,0 +1,23 @@
+export const tableStructure = [
+    {
+      text:'Item',
+      name: 'item',
+      inputType: 'select'
+    },
+    {
+      text:'Material Fee',
+      name: 'materialFee',
+      inputType: 'currency'
+    },  
+    {
+      text:'Packing Fee',
+      name: 'packingFee',
+      inputType: 'currency'
+    },
+    {
+      text:'Unpacking Fee',
+      name: 'unpackingFee',
+      inputType: 'currency'
+    }
+];
+
