+++ src/components/TableBuilder/TableBuilder.scss
@@ -5,7 +5,6 @@
     border-spacing: 0 15px;
     outline: none;
     text-align: left !important;
-    /* background-color:  #eee; */
 }
 
 .Table tr {
