+++ src/components/UI/Select/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './index.css'
+import './index.scss'
 
 export const Select = (props) => {
